removed unnecessary floating point conversion

+++ WBM/Util.cs
@@ -21,7 +21,7 @@ namespace WBM
 
 		public static string formatKDR(int kills, int deaths)
 		{
-			return deaths == 0 ? "inf" : formatDecimal((float)kills / (float)deaths);
+			return deaths == 0 ? "inf" : formatDecimal((float)kills / deaths);
 		}
 
 		public static string formatDecimal(float number)
